+++ driver/src/util.rs
@@ -107,11 +107,10 @@ pub fn batch_processing_state(
 pub fn create_price_finder(fee: Option<Fee>) -> Box<dyn PriceFinding> {
     let optimization_model_string: String =
         env::var("OPTIMIZATION_MODEL").unwrap_or_else(|_| String::from("mip"));
-    let optimization_model = match optimization_model_string.as_str() {
-        "MIP" => OptimizationModel::MIP,
-        "NLP" => OptimizationModel::NLP,
-        _ => OptimizationModel::NAIVE,
-    };
+    // let optimization_model: OptimizationModel =
+    //     OptimizationModel::From(optimization_model_string.clone());
+    let optimization_model = OptimizationModel::from(optimization_model_string.clone());
+
     if optimization_model == OptimizationModel::NAIVE {
         info!("Using naive price finder");
         Box::new(NaiveSolver::new(fee))
